format: add prettierrc (srcs formatter)

+++ backend/nest.js/src/test/test-fake.service.ts
@@ -1,3 +1,4 @@
+/** @prettier */
 import { Inject, Injectable, NotFoundException } from '@nestjs/common';
 import { FriendsService } from 'src/friends/friends.service';
 import { MatchStats } from 'src/game/matchs/entity/matchstats.entity';
@@ -12,7 +13,6 @@ import { SelectQueryBuilder } from 'typeorm/query-builder/SelectQueryBuilder';
 
 @Injectable()
 export class TestFakeService {
-
 	@Inject(UsersService)
 	private readonly usersService: UsersService;
 	@Inject(FriendsService)
@@ -29,50 +29,48 @@ export class TestFakeService {
 		const data = new Array();
 		const allUserIds: number[] = await this.getUsersIds();
 		for (let i: number = 1; i <= nbUsers; ++i) {
-			let fakeUser: {user: User, stats: UserStats, matchs: MatchStats} = await this.createFakeUser(allUserIds)
+			let fakeUser: { user: User; stats: UserStats; matchs: MatchStats } = await this.createFakeUser(allUserIds);
 			data.push(fakeUser);
 			allUserIds.push(fakeUser.user.id);
 		}
 		return data;
 	}
 
-	async createFakeUser(allUserIds: number[]) : Promise<{ user: User, stats: UserStats, matchs: MatchStats}> {
+	async createFakeUser(allUserIds: number[]): Promise<{ user: User; stats: UserStats; matchs: MatchStats }> {
 		const user: User = await this.initUser();
 		const allUserIdsExceptUser: number[] = removeFromArray(allUserIds, user.id);
 
 		const stats: UserStats = await this.initStats(user);
 		const matchs: MatchStats = await this.initMatchHistory(user, allUserIdsExceptUser);
-	
+
 		this.initNewFriendship(user, allUserIdsExceptUser);
-		return ({ user, stats, matchs });
+		return { user, stats, matchs };
 	}
 
-	async addFakeData(user: User): Promise<{ statusCode: number, message: string}> {
+	async addFakeData(user: User): Promise<{ statusCode: number; message: string }> {
 		const allUserIdsExceptUser: number[] = removeFromArray(await this.getUsersIds(), user.id);
 
 		const stats: UserStats = await this.initStats(user);
 		let iMatchs = -1;
 		while (++iMatchs < allUserIdsExceptUser.length) {
 			const matchs: MatchStats = await this.initMatchHistory(user, allUserIdsExceptUser);
-			if (matchs == null)
-				break;
+			if (matchs == null) break;
 		}
 		const allFriendsOfUser: number[] = await this.friendsService.findFriendsIds(user.id);
 		for (let userToRemove of allFriendsOfUser) {
-			removeFromArray(allUserIdsExceptUser, userToRemove)
+			removeFromArray(allUserIdsExceptUser, userToRemove);
 		}
 		let iFriends = -1;
 		while (++iFriends < allUserIdsExceptUser.length) {
 			const target: UserSelectDTO = await this.initNewFriendship(user, allFriendsOfUser);
-			if (!target)
-				break;
+			if (!target) break;
 			removeFromArray(allUserIdsExceptUser, target.id);
 		}
-		return ({ statusCode: 200, message: `${iMatchs} matches and ${iFriends} friend relationships are added.` });
+		return { statusCode: 200, message: `${iMatchs} matches and ${iFriends} friend relationships are added.` };
 	}
 
 	async initUser(): Promise<User> {
-		let user: User = new User;
+		let user: User = new User();
 		user.username = Math.random().toString(36).substring(2, 9);
 		user.login_42 = Math.random().toString(36).substring(2, 9);
 		/*
@@ -80,7 +78,7 @@ export class TestFakeService {
 			return response.url;
 		});
 		*/
-		user.avatar = await toBase64('https://picsum.photos/200')
+		user.avatar = await toBase64('https://picsum.photos/200');
 		user.status = randomEnum(UserStatus);
 		// user.status = UserStatus.IN_GAME;
 		return await this.usersService.add(user);
@@ -102,7 +100,7 @@ export class TestFakeService {
 			return null;
 		}
 		const matchHistory: MatchStats = new MatchStats();
-		
+
 		if (random(0, 2) === 1) {
 			matchHistory.user_id2 = user.id;
 			matchHistory.user_id1 = randomElement(userIds);
@@ -113,10 +111,8 @@ export class TestFakeService {
 
 		const scoreWinner: number = this.randomMaxScoreGame;
 		const scoreLoser: number = random(0, this.randomMaxScoreGame);
-		if (random(0, 2) === 1)
-			matchHistory.score = [scoreWinner, scoreLoser];
-		else
-			matchHistory.score = [scoreLoser, scoreWinner];
+		if (random(0, 2) === 1) matchHistory.score = [scoreWinner, scoreLoser];
+		else matchHistory.score = [scoreLoser, scoreWinner];
 		return this.matchHistoryService.add(matchHistory);
 	}
 
@@ -128,23 +124,21 @@ export class TestFakeService {
 		const randomUserId: number = randomElement(userIds);
 		const randomUser: UserSelectDTO = new UserSelectDTO();
 		randomUser.id = randomUserId;
-		randomUser.username = "fakePlayer";
+		randomUser.username = 'fakePlayer';
 		const randomNb: number = random(1, 4);
 
 		console.log('User', user, 'randomUser', randomUser);
 		await this.friendsService.addFriendRequest(user, randomUser);
-		if (randomNb == 2)
-			await this.friendsService.removeFriendship(randomUser, user);
-		else if (randomNb >= 3)
-			await this.friendsService.acceptFriendRequest(randomUser, user);
+		if (randomNb == 2) await this.friendsService.removeFriendship(randomUser, user);
+		else if (randomNb >= 3) await this.friendsService.acceptFriendRequest(randomUser, user);
 		return randomUser;
 	}
 
 	private async getUsersIds(): Promise<number[]> {
-		const sqlStatement: SelectQueryBuilder<User> = this.usersService.getRepo().createQueryBuilder("user").select(["user.id"]);
+		const sqlStatement: SelectQueryBuilder<User> = this.usersService.getRepo().createQueryBuilder('user').select(['user.id']);
 
 		return await sqlStatement.getMany().then((users: User[]) => {
-			return users.map(u => u.id);
+			return users.map((u) => u.id);
 		}, this.usersService.lambdaDatabaseUnvailable);
 	}
 }
